requirements.txt
boto3~=1.14.52
elasticlogger~=0.1.0
PyYAML~=5.3.1
chalice~=1.18.1